removal of request deps
