+++ api/services/emAnno.ts
@@ -1,55 +1,36 @@
 import * as express from 'express'
-import {config} from '../../../config'
-const generateRequest = require('../../lib/request/request')
-const headerUtilities = require('../../lib/utilities/headerUtilities')
+import { config } from '../../config'
+import { http } from '../lib/http'
+import { getHealth, getInfo } from '../lib/util'
 
 const url = config.services.em_anno_api
 
-function getAnnotionSet(uuid, options) {
-    return generateRequest('GET', `${url}/api/annotation-sets/filter?documentId=${uuid}`, options)
+export async function getAnnotionSet(uuid: string) {
+    const request = await http.get(`${url}/api/annotation-sets/filter?documentId=${uuid}`)
+    return request.data
 }
 
-function createAnnotationSet(options) {
-    return generateRequest('POST', `${url}/api/annotation-sets/`, options)
+export async function createAnnotationSet(body: string) {
+    const request = await http.post(`${url}/api/annotation-sets/`, body)
+    return request.data
+}
+export async function addAnnotation(body: string) {
+    const request = await http.post(`${url}/api/annotations`, body)
+    return request.data
 }
 
-function addAnnotation(options) {
-    return generateRequest('POST', `${url}/api/annotations`, options)
-}
-
-function deleteAnnotation(uuid, options) {
-    return generateRequest('DELETE', `${url}/api/annotations/${uuid}`, options)
-}
-
-function getHealth(options) {
-    return generateRequest('GET', `${url}/health`, options)
-}
-
-function getInfo(options) {
-    return generateRequest('GET', `${url}/info`, options)
-}
-
-function getOptions(req) {
-    return headerUtilities.getAuthHeadersWithBody(req)
+export async function deleteAnnotation(uuid: string) {
+    const request = await http.post(`${url}/api/annotations/${uuid}`)
+    return request.data
 }
 
 module.exports = app => {
     const router = express.Router({ mergeParams: true })
     app.use('/em-anno', router)
 
-    router.get('/health', (req, res, next) => {
-        getHealth(getOptions(req)).pipe(res)
-    })
-
-    router.get('/info', (req, res, next) => {
-        getInfo(getOptions(req)).pipe(res)
-    })
-
     router.post('/annotation-sets', (req, res, next) => {
         // Called when get annotation-sets returns 404
-        const options = getOptions(req)
-
-        createAnnotationSet(options)
+        createAnnotationSet(req.body)
             .then(response => {
                 res.setHeader('Access-Control-Allow-Origin', '*')
                 res.setHeader('content-type', 'application/json')
@@ -61,10 +42,9 @@ module.exports = app => {
     })
 
     router.get('/annotation-sets/:uuid', (req, res, next) => {
-        const options = getOptions(req)
         const uuid = req.params.uuid
 
-        getAnnotionSet(uuid, options)
+        getAnnotionSet(uuid)
             .then(response => {
                 res.setHeader('Access-Control-Allow-Origin', '*')
                 res.setHeader('content-type', 'application/json')
@@ -77,10 +57,9 @@ module.exports = app => {
     })
 
     router.delete('/annotations/:uuid', (req, res, next) => {
-        const options = getOptions(req)
         const uuid = req.params.uuid
 
-        deleteAnnotation(uuid, options)
+        deleteAnnotation(uuid)
             .then(response => {
                 res.setHeader('Access-Control-Allow-Origin', '*')
                 res.setHeader('content-type', 'application/json')
@@ -92,9 +71,8 @@ module.exports = app => {
     })
 
     router.post('/annotations', (req, res, next) => {
-        const options = getOptions(req)
 
-        addAnnotation(options)
+        addAnnotation(req.body)
             .then(response => {
                 res.setHeader('Access-Control-Allow-Origin', '*')
                 res.setHeader('content-type', 'application/json')
@@ -106,14 +84,4 @@ module.exports = app => {
     })
 }
 
-module.exports.getInfo = getInfo
 
-module.exports.getHealth = getHealth
-
-module.exports.getAnnotionSet = getAnnotionSet
-
-module.exports.createAnnotationSet = createAnnotationSet
-
-module.exports.addAnnotation = addAnnotation
-
-module.exports.deleteAnnotation = deleteAnnotation
