+++ package.json
@@ -100,9 +100,6 @@
     "otp": "0.1.3",
     "p-iteration": "^1.1.7",
     "pdfjs-dist": "2.0.943",
-    "request": "2.88.0",
-    "request-promise": "4.2.2",
-    "request-promise-native": "1.0.7",
     "rxjs": "6.4.0",
     "rxjs-compat": "6.4.0",
     "session-file-store": "1.2.0",
